support snapshot and images creation in cells
