+++ cloud/disk_manager/internal/pkg/facade/image_service_nemesis_test/ya.make
@@ -8,6 +8,8 @@ INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common
 
 GO_XTEST_SRCS(
     ../image_service_test/image_service_test.go
+    ../image_service_test/common_test.go
+    ../image_service_test/image_service_cells_test.go
 )
 
 DEPENDS(
