+++ cloud/disk_manager/internal/pkg/facade/image_service_test/image_service_cells_test.go
@@ -1,0 +1,19 @@
+package tests
+
+import (
+	"testing"
+
+	disk_manager "github.com/ydb-platform/nbs/cloud/disk_manager/api"
+)
+
+////////////////////////////////////////////////////////////////////////////////
+
+func TestImageServiceCellsCreateImageFromDisk(t *testing.T) {
+	testImageServiceCreateImageFromDiskWithKind(
+		t,
+		disk_manager.DiskKind_DISK_KIND_SSD,
+		uint64(4194304),
+		shardedZoneID,
+		cellID, // diskCellID. Zone is sharded. Policy is FIRST_IN_CONFIG.
+	)
+}
