+++ cloud/disk_manager/internal/pkg/facade/image_service_test/image_service_test.go
@@ -176,98 +176,6 @@ func checkUnencryptedImage(
 	require.ErrorContains(t, err, "KeyPath should contain path to encryption key")
 }
 
-func testImageServiceCreateImageFromDiskWithKind(
-	t *testing.T,
-	diskKind disk_manager.DiskKind,
-	diskSize uint64,
-) {
-
-	ctx := testcommon.NewContext()
-
-	client, err := testcommon.NewClient(ctx)
-	require.NoError(t, err)
-	defer client.Close()
-
-	diskID := t.Name()
-
-	reqCtx := testcommon.GetRequestContext(t, ctx)
-	operation, err := client.CreateDisk(reqCtx, &disk_manager.CreateDiskRequest{
-		Src: &disk_manager.CreateDiskRequest_SrcEmpty{
-			SrcEmpty: &empty.Empty{},
-		},
-		Size: int64(diskSize),
-		Kind: diskKind,
-		DiskId: &disk_manager.DiskId{
-			ZoneId: "zone-a",
-			DiskId: diskID,
-		},
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, operation)
-	err = internal_client.WaitOperation(ctx, client, operation.Id)
-	require.NoError(t, err)
-
-	nbsClient := testcommon.NewNbsTestingClient(t, ctx, "zone-a")
-	diskContentInfo, err := nbsClient.FillDisk(ctx, diskID, diskSize)
-	require.NoError(t, err)
-
-	imageID := t.Name()
-
-	reqCtx = testcommon.GetRequestContext(t, ctx)
-	operation, err = client.CreateImage(reqCtx, &disk_manager.CreateImageRequest{
-		Src: &disk_manager.CreateImageRequest_SrcDiskId{
-			SrcDiskId: &disk_manager.DiskId{
-				ZoneId: "zone-a",
-				DiskId: diskID,
-			},
-		},
-		DstImageId: imageID,
-		FolderId:   "folder",
-		Pooled:     true,
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, operation)
-
-	response := disk_manager.CreateImageResponse{}
-	err = internal_client.WaitResponse(ctx, client, operation.Id, &response)
-	require.NoError(t, err)
-	require.Equal(t, int64(diskSize), response.Size)
-
-	meta := disk_manager.CreateImageMetadata{}
-	err = internal_client.GetOperationMetadata(ctx, client, operation.Id, &meta)
-	require.NoError(t, err)
-	require.Equal(t, float64(1), meta.Progress)
-
-	diskParams, err := nbsClient.Describe(ctx, diskID)
-	require.NoError(t, err)
-
-	if diskParams.IsDiskRegistryBasedDisk {
-		testcommon.RequireCheckpointsDoNotExist(t, ctx, diskID)
-	} else {
-		testcommon.RequireCheckpoint(t, ctx, diskID, imageID)
-	}
-
-	checkUnencryptedImage(
-		t,
-		client,
-		ctx,
-		imageID,
-		int64(diskSize),
-		diskContentInfo.Crc32,
-	)
-
-	reqCtx = testcommon.GetRequestContext(t, ctx)
-	operation, err = client.DeleteImage(reqCtx, &disk_manager.DeleteImageRequest{
-		ImageId: imageID,
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, operation)
-	err = internal_client.WaitOperation(ctx, client, operation.Id)
-	require.NoError(t, err)
-
-	testcommon.CheckConsistency(t, ctx)
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 
 func TestImageServiceCreateImageFromImage(t *testing.T) {
@@ -1002,6 +910,8 @@ func TestImageServiceCreateImageFromDisk(t *testing.T) {
 		t,
 		disk_manager.DiskKind_DISK_KIND_SSD,
 		uint64(4194304),
+		defaultZoneID,
+		defaultZoneID, // diskCellID. Zone is not sharded.
 	)
 }
 
@@ -1010,6 +920,8 @@ func TestImageServiceCreateImageFromNonReplicatedDisk(t *testing.T) {
 		t,
 		disk_manager.DiskKind_DISK_KIND_SSD_NONREPLICATED,
 		uint64(1073741824),
+		defaultZoneID,
+		defaultZoneID, // diskCellID. Zone is not sharded.
 	)
 }
