+++ cloud/disk_manager/internal/pkg/facade/snapshot_service_test/snapshot_service_test.go
@@ -18,121 +18,47 @@ import (
 
 ////////////////////////////////////////////////////////////////////////////////
 
-func testCreateSnapshotFromDisk(
-	t *testing.T,
-	diskKind disk_manager.DiskKind,
-	diskBlockSize uint32,
-	diskSize uint64,
-) {
-
-	ctx := testcommon.NewContext()
-
-	client, err := testcommon.NewClient(ctx)
-	require.NoError(t, err)
-	defer client.Close()
-
-	diskID := t.Name()
-
-	reqCtx := testcommon.GetRequestContext(t, ctx)
-	operation, err := client.CreateDisk(reqCtx, &disk_manager.CreateDiskRequest{
-		Src: &disk_manager.CreateDiskRequest_SrcEmpty{
-			SrcEmpty: &empty.Empty{},
-		},
-		Size: int64(diskSize),
-		Kind: diskKind,
-		DiskId: &disk_manager.DiskId{
-			ZoneId: "zone-a",
-			DiskId: diskID,
-		},
-		BlockSize: int64(diskBlockSize),
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, operation)
-	err = internal_client.WaitOperation(ctx, client, operation.Id)
-	require.NoError(t, err)
-
-	nbsClient := testcommon.NewNbsTestingClient(t, ctx, "zone-a")
-	_, err = nbsClient.FillDisk(ctx, diskID, 64*4096)
-	require.NoError(t, err)
-
-	snapshotID := t.Name()
-
-	reqCtx = testcommon.GetRequestContext(t, ctx)
-	operation, err = client.CreateSnapshot(reqCtx, &disk_manager.CreateSnapshotRequest{
-		Src: &disk_manager.DiskId{
-			ZoneId: "zone-a",
-			DiskId: diskID,
-		},
-		SnapshotId: snapshotID,
-		FolderId:   "folder",
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, operation)
-
-	response := disk_manager.CreateSnapshotResponse{}
-	err = internal_client.WaitResponse(ctx, client, operation.Id, &response)
-	require.NoError(t, err)
-	require.Equal(t, int64(diskSize), response.Size)
-
-	meta := disk_manager.CreateSnapshotMetadata{}
-	err = internal_client.GetOperationMetadata(ctx, client, operation.Id, &meta)
-	require.NoError(t, err)
-	require.Equal(t, float64(1), meta.Progress)
-
-	diskParams, err := nbsClient.Describe(ctx, diskID)
-	require.NoError(t, err)
-
-	if diskParams.IsDiskRegistryBasedDisk {
-		testcommon.RequireCheckpointsDoNotExist(t, ctx, diskID)
-	} else {
-		testcommon.RequireCheckpoint(t, ctx, diskID, snapshotID)
-	}
-
-	reqCtx = testcommon.GetRequestContext(t, ctx)
-	operation, err = client.DeleteSnapshot(reqCtx, &disk_manager.DeleteSnapshotRequest{
-		SnapshotId: snapshotID,
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, operation)
-	err = internal_client.WaitOperation(ctx, client, operation.Id)
-	require.NoError(t, err)
-
-	testcommon.CheckConsistency(t, ctx)
-}
-
 func TestSnapshotServiceCreateSnapshotFromDisk(t *testing.T) {
-	testCreateSnapshotFromDisk(
+	testCreateSnapshotFromDiskWithZoneID(
 		t,
 		disk_manager.DiskKind_DISK_KIND_SSD,
 		4096,         // diskBlockSize
 		32*1024*4096, // diskSize
+		defaultZoneID,
+		defaultZoneID, // diskCellID. Zone is not sharded.
 	)
 }
 
 func TestSnapshotServiceCreateSnapshotFromLargeDisk(t *testing.T) {
-	testCreateSnapshotFromDisk(
+	testCreateSnapshotFromDiskWithZoneID(
 		t,
 		disk_manager.DiskKind_DISK_KIND_SSD,
 		65536,         // diskBlockSize
 		1000000*65536, // diskSize
+		defaultZoneID,
+		defaultZoneID, // diskCellID. Zone is not sharded.
 	)
 }
 
 func TestSnapshotServiceCreateSnapshotFromSsdNonreplicatedDisk(t *testing.T) {
-	testCreateSnapshotFromDisk(
+	testCreateSnapshotFromDiskWithZoneID(
 		t,
 		disk_manager.DiskKind_DISK_KIND_SSD_NONREPLICATED,
 		4096,        // diskBlockSize
 		262144*4096, // diskSize
+		defaultZoneID,
+		defaultZoneID, // diskCellID. Zone is not sharded.
 	)
 }
 
 func TestSnapshotServiceCreateSnapshotFromHddNonreplicatedDisk(t *testing.T) {
-	testCreateSnapshotFromDisk(
+	testCreateSnapshotFromDiskWithZoneID(
 		t,
 		disk_manager.DiskKind_DISK_KIND_HDD_NONREPLICATED,
 		4096,        // diskBlockSize
 		262144*4096, // diskSize
+		defaultZoneID,
+		defaultZoneID, // diskCellID. Zone is not sharded.
 	)
 }
