+++ cloud/disk_manager/internal/pkg/facade/snapshot_service_test/ya.make
@@ -5,6 +5,8 @@ SET_APPEND(RECIPE_ARGS --disk-agent-count 3)
 INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
 
 GO_XTEST_SRCS(
+    common_test.go
+    snapshot_service_cells_test.go
     snapshot_service_test.go
 )
